requirements.txt
absl-py==1.4.0
anyio==3.7.0
apex @ file:///home/jh-raph-park/workspace/git_repo/apex
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.9.1
bleach==6.0.0
catkin-pkg==0.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.26.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
dm-control==1.0.11
dm-env==1.6
dm-tree==0.1.8
docutils==0.19
einops==0.6.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.10.7
fonttools==4.39.3
fqdn==1.5.1
functorch==2.0.0
gin-config==0.5.0
glfw==2.5.7
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
h5py-cache==1.0
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.2
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
labmaze==1.0.6
lit==16.0.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
mujoco==2.3.3
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
overrides==7.3.1
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.6.0
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.10.4
pycparser==2.21
Pygments==2.14.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
pytest==7.3.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rospkg==1.5.0
scikit-learn==1.3.0
scikit-sparse==0.4.8
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
terminado==0.17.1
-e git+https://github.com/facebookresearch/theseus.git@016c1f71ac1e785124a8716b304c85586079f9bb#egg=theseus_ai
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
zipp==3.15.0
